requirements.txt
annotated-types==0.7.0
anyio==4.11.0
branca==0.8.2
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.118.3
feature_engine==1.9.3
folium==0.20.0
fonttools==4.60.1
h11==0.16.0
idna==3.10
imbalanced-learn==0.14.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lightgbm==4.6.0
MarkupSafe==3.0.3
matplotlib==3.10.7
narwhals==2.7.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
patsy==1.0.1
pillow==11.3.0
plotly==6.3.1
pydantic==2.12.0
pydantic_core==2.41.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
statsmodels==0.14.5
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
xyzservices==2025.4.0
